Move graphiql to helpers
