+++ src/helpers/index.js
@@ -1,5 +1,6 @@
 'use strict';
 
 module.exports = {
+  graphiql: require('./graphiql.js'),
   expressGraphqlExtension: require('./express-graphql-extension.js')
 };
