+++ test/index.test.js
@@ -5,9 +5,11 @@ const test = require('tape');
 const index = require('../index.js');
 
 test('index: all functions successfully imported', function (t) {
-  ['createClient', 'prepareSpaceGraph', 'createSchema', 'createUI'].forEach(m => {
+  ['createClient', 'prepareSpaceGraph', 'createSchema'].forEach(m => {
     t.equal(typeof index[m], 'function');
   });
+
+  t.equal(typeof index.helpers, 'object');
 
   t.end();
 });
